@@ -13,45 +13,47 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "^16.1.4",
-    "@angular/core": "^16.1.4",
-    "@angular/forms": "^16.1.4",
-    "@angular/platform-browser": "^16.1.4",
-    "@angular/platform-browser-dynamic": "^16.1.4",
-    "@angular/router": "^16.1.4",
-    "@capacitor-community/sqlite": "^4.2.2",
-    "@capacitor/app": "latest",
-    "@capacitor/core": "latest",
-    "@capacitor/haptics": "latest",
-    "@capacitor/ios": "latest",
-    "@capacitor/keyboard": "latest",
-    "@capacitor/status-bar": "latest",
-    "@ionic/angular": "^7.1.2",
-    "@ionic/pwa-elements": "^3.1.1",
+    "@angular/cdk": "^16.2.10",
+    "@angular/common": "^16.2.10",
+    "@angular/core": "^16.2.10",
+    "@angular/forms": "^16.2.10",
+    "@angular/platform-browser": "^16.2.10",
+    "@angular/platform-browser-dynamic": "^16.2.10",
+    "@angular/router": "^16.2.10",
+    "@capacitor-community/sqlite": "^5.4.1",
+    "@capacitor/app": "^5.0.6",
+    "@capacitor/core": "^5.5.0",
+    "@capacitor/haptics": "^5.0.6",
+    "@capacitor/ios": "^5.5.0",
+    "@capacitor/keyboard": "^5.0.6",
+    "@capacitor/status-bar": "^5.0.6",
+    "@ionic/angular": "^7.5.1",
+    "@ionic/pwa-elements": "^3.2.2",
     "@ngx-translate/core": "^14.0.0",
     "@ngx-translate/http-loader": "^7.0.0",
     "capitalize": "^2.0.4",
-    "jeep-sqlite": "^1.6.7",
+    "jeep-sqlite": "^1.7.1",
     "rxjs": "^7.5.0",
-    "swiper": "^10.0.4",
-    "tslib": "^2.2.0",
-    "zone.js": "^0.13.1"
+    "swiper": "^10.3.1",
+    "ts-md5": "^1.3.1",
+    "tslib": "^2.6.2",
+    "zone.js": "^0.13.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^16.1.4",
+    "@angular-devkit/build-angular": "^16.2.7",
     "@angular-devkit/core": "^16.1.4",
     "@angular-devkit/schematics": "^16.1.4",
     "@angular-eslint/builder": "~13.0.1",
     "@angular-eslint/eslint-plugin": "~13.0.1",
     "@angular-eslint/eslint-plugin-template": "~13.0.1",
     "@angular-eslint/template-parser": "~13.0.1",
-    "@angular/cli": "^16.1.4",
-    "@angular/compiler": "^16.1.4",
-    "@angular/compiler-cli": "^16.1.4",
-    "@angular/language-service": "^16.1.4",
-    "@capacitor/cli": "latest",
+    "@angular/cli": "^16.2.7",
+    "@angular/compiler": "^16.2.10",
+    "@angular/compiler-cli": "^16.2.10",
+    "@angular/language-service": "^16.2.10",
+    "@capacitor/cli": "^5.5.0",
     "@ionic/angular-toolkit": "^9.0.0",
-    "@schematics/angular": "^14.2.7",
+    "@schematics/angular": "^14.2.13",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.20.55",
